hard work

+++ utils/3dlut.h
@@ -7,7 +7,7 @@ class Lookup_table{
 	bool m_is3d;
 	float m_domain_min[3];
 	float m_domain_max[3];
-	unsigned int m_dimension = 0;
+	unsigned short m_dimension = 0;
 public:
 	Lookup_table(std::string filename);
 	~Lookup_table();
